feat: add context manager for setting TensorFlow's random sstate
